refactor(creative_agent): split tools.py into image/bq/gcs modules with re-exports
